사이드바 코드 리펙터링
