+++ src/components/Layout/UserCurrentSituation/index.tsx
@@ -1,6 +1,6 @@
 import CurrentCardContainer from "@/components/Composition/CurrentCardContainer";
 import React from "react";
-import { dashboardPageName } from "../AfterLoginLoyouts/index.const";
+import { dashboardPageName } from "../AfterLoginLayouts/index.const";
 
 const UserCurrentSituation = () => {
   return (
