+++ src/styles/globals.css
@@ -48,3 +48,7 @@ body {
 .custom-page-name {
   @apply font-bold text-3xl mb-2;
 }
+
+.basicSidbarbutton {
+  @apply w-48 px-2 py-3 rounded-md
+}
